requirements.txt
pandas>=2.2.3
numpy>=1.26.4
scikit-learn==1.7.2
transformers==4.57.0
stable-baselines3>=2.3.2
gymnasium>=0.29.1
joblib>=1.4.2
ib_insync>=0.9.70
httpx>=0.27.0
dash>=2.17.1
plotly>=5.23.0
pytest>=8.3.2
retry>=0.9.2
python-dotenv>= 1.2.1
psutil>=6.0.0
schedule>=1.1.0
watchdog>=4.0.0
imbalanced-learn>=0.12.3
optuna>=4.6.0
tqdm>=4.67.1
lightgbm>=4.6.0
xgboost>=3.1.1
